requirements.txt
anyio==3.6.2
attrs==19.3.0
Automat==0.8.0
autopep8==2.0.1
black==22.12.0
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
click==8.1.3
cloud-init==20.1
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
dbus-python==1.2.16
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==2.2.1
email-validator==1.3.0
entrypoints==0.3
fastapi==0.88.0
greenlet==2.0.1
h11==0.14.0
httpcore==0.16.3
httplib2==0.14.0
httptools==0.5.0
httpx==0.23.1
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupSafe==2.1.1
more-itertools==4.2.0
mypy-extensions==0.4.3
mysql-connector-python==8.0.31
netifaces==0.10.4
oauthlib==3.1.0
orjson==3.8.3
pathspec==0.10.3
platformdirs==2.6.2
protobuf==3.20.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycodestyle==2.10.0
pydantic==1.10.4
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0
python-debian===0.1.36ubuntu1
python-dotenv==0.21.0
python-multipart==0.0.5
PyYAML==5.3.1
requests==2.22.0
requests-unixsocket==0.2.0
rfc3986==1.5.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sniffio==1.3.0
SQLAlchemy==1.4.45
ssh-import-id==5.10
starlette==0.22.0
systemd-python==234
tomli==2.0.1
Twisted==18.9.0
typing-extensions==4.4.0
ubuntu-advantage-tools==20.3
ufw==0.36
ujson==5.6.0
unattended-upgrades==0.1
urllib3==1.25.8
uvicorn==0.20.0
uvloop==0.17.0
wadllib==1.3.3
watchfiles==0.18.1
websockets==10.4
zipp==1.0.0
zope.interface==4.7.1
